requirements.txt
# Pairs Trading System Requirements
# IC Markets Global via MetaTrader 5

# Core
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Statistical Analysis
statsmodels>=0.14.0

# MetaTrader 5 Integration
MetaTrader5>=5.0.45

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0

# Visualization (optional)
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Data Storage
pyarrow>=12.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0
